refactor(player): extract shared core into @gbfm/player
Move the platform-agnostic player logic out of apps/mobile into a shared
package so web can build on the same core: queue reducer, playback state
machines, persisted-queue schema, storage factory, and play delivery.

The storage factory already carried a localStorage adapter for expo-web,
so the platform seam existed; it just lived in the mobile app. Mobile now
keeps only its Expo bindings (FileSystem adapter, expo-audio subscription)
and imports the rest.

No behavior change. Tests move with the code.
