+++ package.json
@@ -11,7 +11,7 @@
     "gen": "cd vps && sst shell drizzle-kit generate",
     "release": "semantic-release",
     "typecheck": "tsgo --noEmit && bun --filter='*' typecheck",
-    "test": "cd packages/core && bun run test && cd ../theme && bun run test && cd ../../apps/mobile && bun run test && cd ../vps && bun run test",
+    "test": "cd packages/core && bun run test && cd ../theme && bun run test && cd ../player && bun run test && cd ../../apps/mobile && bun run test && cd ../vps && bun run test",
     "lint": "oxlint apps/*/src packages infra '*.{js,ts,tsx,jsx}' --no-error-on-unmatched-pattern",
     "lint:fix": "oxlint apps/*/src packages infra '*.{js,ts,tsx,jsx}' --fix --no-error-on-unmatched-pattern",
     "format": "oxfmt apps/*/src packages infra '*.{js,ts,tsx,jsx,json,css}' --write --no-error-on-unmatched-pattern",
