+++ packages/player/tsconfig.json
@@ -1,0 +1,20 @@
+{
+  "compilerOptions": {
+    "target": "ES2022",
+    "module": "ESNext",
+    "lib": ["ES2022"],
+    "moduleResolution": "bundler",
+    "strict": true,
+    "esModuleInterop": true,
+    "skipLibCheck": true,
+    "resolveJsonModule": true,
+    "isolatedModules": true,
+    "noEmit": true,
+    "declaration": true,
+    "declarationMap": true,
+    "types": ["@types/bun"],
+    "plugins": [{ "name": "@effect/language-service" }]
+  },
+  "include": ["src/**/*"],
+  "exclude": ["node_modules", "dist"]
+}
